News Application using MVC DAO pattern
This java application gives the user to choose one database option to choose from two available options. LinkedHashMap is used as a temporary storage for one type and MongoDB is used as the other database. This is a multi-role web application having three types of users namely: Guest, Subscriber and Reporter.